CASMTRIAGE-3786: be more careful when deleting .ssh/config
The cleanup function referenced by `trap` in prerequisites.sh
is also used in other contexts. However, only prerequisites.sh
deals with creating a backup ssh config file.  If the trap function
is run from another context, it blindly removes the system ssh
config.

+++ upgrade/1.2/scripts/common/upgrade-state.sh
@@ -85,8 +85,10 @@ function err_report() {
     echo "${cmd}"
 
     # restore previous ssh config if there was one, remove ours
-    rm -f /root/.ssh/config
-    test -f /root/.ssh/config.bak && mv /root/.ssh/config.bak /root/.ssh/config
+    if [[ $RESTORE_SSH_CONFIG -eq 1 ]] ; then
+        rm -f /root/.ssh/config
+        test -f /root/.ssh/config.bak && mv /root/.ssh/config.bak /root/.ssh/config
+    fi
 
     # ignore some internal expected errors
     local ignoreCmd="cray artifacts list config-data"
